+++ packages/core/src/component/component.ts
@@ -2,32 +2,32 @@ import type { TSESTreeClass, TSESTreeFunction } from "@eslint-react/ast";
 import type { O } from "@eslint-react/tools";
 import type { TSESTree } from "@typescript-eslint/types";
 
-import type { ExRComponentCollectorHint } from "./component-collector-hint";
-import type { ExRClassComponentFlag, ExRFunctionComponentFlag } from "./component-flag";
-import type { ExRComponentInitPath } from "./component-init-path";
+import type { ERComponentCollectorHint } from "./component-collector-hint";
+import type { ERClassComponentFlag, ERFunctionComponentFlag } from "./component-flag";
+import type { ERComponentInitPath } from "./component-init-path";
 
-export type ExRFunctionComponent = {
+export type ERFunctionComponent = {
   _: string;
   id: O.Option<TSESTree.Identifier | TSESTree.Identifier[]>;
   kind: "function";
   node: TSESTreeFunction;
   name: O.Option<string>;
-  flag: ExRFunctionComponentFlag;
-  hint: ExRComponentCollectorHint;
-  initPath: O.Option<ExRComponentInitPath>;
+  flag: ERFunctionComponentFlag;
+  hint: ERComponentCollectorHint;
+  initPath: O.Option<ERComponentInitPath>;
   hookCalls: TSESTree.CallExpression[];
   displayName: O.Option<TSESTree.Expression>;
 };
 
-export type ExRClassComponent = {
+export type ERClassComponent = {
   _: string;
   id: O.Option<TSESTree.Identifier>;
   kind: "class";
   node: TSESTreeClass;
   name: O.Option<string>;
-  flag: ExRClassComponentFlag;
+  flag: ERClassComponentFlag;
   methods: (TSESTree.MethodDefinition | TSESTree.PropertyDefinition)[];
   displayName: O.Option<TSESTree.Expression>;
 };
 
-export type ExRComponent = ExRClassComponent | ExRFunctionComponent;
+export type ERComponent = ERClassComponent | ERFunctionComponent;