+++ packages/types/src/settings.ts
@@ -8,7 +8,7 @@ export type ReactSettings = ReadonlyDeep<{
 
 // The `settings` object in eslint config for all plugins.
 // We only care about the `react` field at the moment.
-export type ExRSettings = ReadonlyDeep<{
+export type ERSettings = ReadonlyDeep<{
   [key: string]: unknown;
   react: ReactSettings;
 }>;